Refactor translation loading
Instead of loading and parsing CSV files directly, the app will now use
the `.translation` resource files created by Godot when importing a CSV
with translations. This simplifies the `TranslationLoader` script
significantly and removes a potential inconsistency with CSV parsing.

During the change it became clear that `Translation.get_message_ids()`
will always return an empty array for translations from CSV files. The
method was used to determine if a translation contains a message. This
process has been changed to just use the `Translation.get_message()`
method and check if the resulting string is empty.

As an added bonus, this made the analyzing process _much_ faster, by a
factor of 10.