Reduce target fps on focus loss
This change is created to fix a weird issue with Godot unlocking FPS
limits in debugging mode, when the window is in the background (i.e. not
visible). For me this results in framerates in the hundreds, causing
high load and fans to spool up.

To fix this, the app script listens for the NOTIFICATION_WM_FOCUS_OUT
and WM_NOTIFICATION_FOCUS_IN notifications and will set the target FPS
to 60 when focus is lost. When focus is regained, the fps target will be
set to 0 i.e. synced to the display refresh rate.