Remove initial values from _ready
With the recent change of not deferring the load_state call in a state's
_ready() callback, a state could override persisted properties by
calling set_state in _ready(). This only occured in AppState, so the
script has been changed to initialize its properties with the default
values.

Going forward, if a state wants to set default values for persisted
properties, it should do one of the following:

- Initialize the property with the default value upon declaration.
- Declare an _init() callback and call set_state() there.
- Set the `immediate` param of set_state() to `true` in _ready()
callbacks.